requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.0
category-encoders==2.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudscraper==1.2.69
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
data-describe==0.1.0b3
db-dtypes==1.1.1
debugpy==1.6.7
decorator==5.1.1
deepchecks==0.13.1
defusedxml==0.7.1
exceptiongroup==1.1.1
executing==1.2.0
fake-useragent==1.1.3
fastjsonschema==2.16.3
fonttools==4.39.3
gensim==4.3.1
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-bigquery==3.9.0
google-cloud-core==2.3.2
google-cloud-secret-manager==2.16.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.53.0
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.5
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==3.0.1
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter-server==1.23.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.4
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.0.1
platformdirs==3.2.0
plotly==5.14.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
PyNomaly==0.3.3
pyparsing==3.0.9
pyppeteer-stealth==2.7.4
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-utils==3.5.2
pytz==2023.3
pyzmq==23.2.1
requests==2.28.2
requests-toolbelt==0.10.1
retry==0.9.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.8.3
Send2Trash==1.8.0
setuptools==65.6.3
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sodapy==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4
stack-data==0.6.2
statsmodels==0.13.5
sweetviz==2.1.4
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.5.1
wheel==0.38.4
widgetsnbextension==3.6.4
wsproto==1.2.0
